% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{div}
\alias{div}
\title{División entre dos números}
\usage{
div(a, b)
}
\arguments{
\item{a}{Primer número}

\item{b}{Segundo número}
}
\value{
La división entre a y b
}
\description{
Esta función devuelve la división entre dos números.
}
\examples{
div(3,7.2)

}
\seealso{
\link{mult}, \link{suma}, \link{resta}.
}
