% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult.R
\name{mult}
\alias{mult}
\title{Multiplicación entre dos números}
\usage{
mult(a, b)
}
\arguments{
\item{a}{Primer número}

\item{b}{Segundo número}
}
\value{
El producto entre a y b
}
\description{
Esta función devuelve la multiplicación entre dos números.
}
\examples{
mult(4,2.3)
mult(0,3.7)
mult(-4/3,98.123)
mult(pi,3)

}
\seealso{
\link{div}, \link{suma}, \link{resta}.
}
